Use byte arrays as standard

+++ cbits/simd_avx2.c
@@ -33,7 +33,7 @@ void avx2_memcpy(
 
 void avx2_cmpeq8(
     uint8_t byte,
-    uint64_t *target,
+    uint8_t *target,
     size_t target_length,
     uint8_t *source) {
 #if defined(AVX2_ENABLED)
